requirements.txt
Pillow>=10.0.0
ExifRead>=3.0.0
pyproj>=3.6.0
ifcopenshell>=0.7.0
numpy>=1.24.0
